+++ app/services/emails/EmailDefinitionsConsumer.scala
@@ -314,6 +314,40 @@ object EmailDefinitionsConsumer {
     }
 
   }
+  case object ConsumerProResponseNotificationOnAdminCompletion extends EmailDefinition {
+    override val category = Consumer
+
+    override def examples =
+      Seq(
+        "report_ack_pro_consumer_on_admin_completion" -> ((recipient, messagesApi) =>
+          EmailImpl(genReport.copy(email = recipient), Some(genCompany), messagesApi)
+        )
+      )
+
+    final case class EmailImpl(
+        report: Report,
+        maybeCompany: Option[Company],
+        messagesApi: MessagesApi
+    ) extends ConsumerEmail {
+      private val lang                                        = Lang(getLocaleOrDefault(report.lang))
+      implicit private val messagesProvider: MessagesProvider = MessagesImpl(lang, messagesApi)
+
+      override val recipients: List[EmailAddress] = List(report.email)
+      override val subject: String = messagesApi("ConsumerReportAckProEmailOnAdminCompletion.subject")(lang)
+
+      override def getBody: (FrontRoute, EmailAddress) => String = (_, _) =>
+        views.html.mails.consumer
+          .reportToConsumerAcknowledgmentOnAdminCompletion(
+            report,
+            maybeCompany
+          )
+          .toString
+
+      override def getAttachements: AttachmentService => Seq[Attachment] =
+        s => s.defaultAttachments ++ s.attachmentSeqForWorkflowStepN(4, report.lang.getOrElse(Locale.FRENCH))
+    }
+
+  }
 
   private def getLocaleOrDefault(locale: Option[Locale]): Locale = locale.getOrElse(Locale.FRENCH)
